// camm-saas-fe/src/Models/Company/CompanyInterface.ts
export interface ICompany {
	id?: number;
	name?: string;
	firstAddress?: string;
	secondAddress?: string;
	thirdAddress?: string;
	city?: string;
	province?: string;
	country?: string;
	postalCode?: string;
	latitude?: number;
	longitude?: number;
	biography?: string;
	twitter?: string;
	instagram?: string;
	facebook?: string;
	linkedin?: string;
	mainImage?: string;
	contactEmail?: string;
	isMember?: boolean;
}